Better forced mode indication
